proyectos/p02-agente-multi-herramienta/requirements.txt
streamlit==1.35.0
groq==0.9.0
tavily-python==0.3.3
python-dotenv==1.0.1
pymupdf==1.24.2
langchain-text-splitters==0.2.0
python-docx==1.1.0  # Para exportar informes profesionales
langchain-text-splitters==0.2.0 # Para manejar PDFs largos